requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
pydantic==1.10.7
pyodbc==4.0.34
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3